Separate math module from utils
